Enable Solhint (#10869)
* enable solhint for /shared/ folder

add others to ignore file
add solhint to CICD

* add chainlink-solidity ruleset

* reduce ignore scope & fix interface folder issues

* rm unused imports

* clean /libraries

* set no-unused-import to error

* allow 968 warnings, the number we currently have

* fix second layer imports

* contracts/v0.8: fix solhint warnings for vrf contracts (#10875)

* rm no-inline-assembly rule

---------

+++ contracts/src/v0.8/dev/vrf/BlockhashStore.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.6;
 
-import "../../ChainSpecificUtil.sol";
+import {ChainSpecificUtil} from "../../ChainSpecificUtil.sol";
 
 /**
  * @title BlockhashStore
@@ -14,7 +14,7 @@ import "../../ChainSpecificUtil.sol";
  *   would have to be deployed.
  */
 contract BlockhashStore {
-  mapping(uint => bytes32) internal s_blockhashes;
+  mapping(uint256 => bytes32) internal s_blockhashes;
 
   /**
    * @notice stores blockhash of a given block, assuming it is available through BLOCKHASH
@@ -22,6 +22,7 @@ contract BlockhashStore {
    */
   function store(uint256 n) public {
     bytes32 h = ChainSpecificUtil.getBlockhash(uint64(n));
+    // solhint-disable-next-line custom-errors
     require(h != 0x0, "blockhash(n) failed");
     s_blockhashes[n] = h;
   }
@@ -40,6 +41,7 @@ contract BlockhashStore {
    *   that it hashes to a stored blockhash, and then extract parentHash to get the n-th blockhash.
    */
   function storeVerifyHeader(uint256 n, bytes memory header) public {
+    // solhint-disable-next-line custom-errors
     require(keccak256(header) == s_blockhashes[n + 1], "header has unknown blockhash");
 
     // At this point, we know that header is the correct blockheader for block n+1.
@@ -72,6 +74,7 @@ contract BlockhashStore {
    */
   function getBlockhash(uint256 n) external view returns (bytes32) {
     bytes32 h = s_blockhashes[n];
+    // solhint-disable-next-line custom-errors
     require(h != 0x0, "blockhash not found in store");
     return h;
   }
